classification_supervise/submeeting2022/code/requirements.txt
albumentations==1.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
imageio==2.25.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python-headless==4.7.0.68
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2023.2.3
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.14
